Refactor entities file structure a bit
