+++ app/src/main/java/net/bible/android/control/page/CurrentBibleVerse.kt
@@ -18,13 +18,11 @@
 package net.bible.android.control.page
 
 import net.bible.android.control.versification.ConvertibleVerse
-import net.bible.android.database.workspaces.WorkspaceEntities
+import net.bible.android.database.WorkspaceEntities
 import org.crosswire.jsword.passage.Verse
 import org.crosswire.jsword.versification.BibleBook
 import org.crosswire.jsword.versification.Versification
 import org.crosswire.jsword.versification.system.Versifications
-import org.json.JSONException
-import org.json.JSONObject
 
 /**
  * @author Martin Denham [mjdenham at gmail dot com]
