+++ app/src/main/java/net/bible/android/control/page/CurrentGeneralBookPage.kt
@@ -21,7 +21,6 @@ import android.app.Activity
 import android.view.Menu
 import net.bible.android.activity.R
 import net.bible.android.view.activity.navigation.genbookmap.ChooseGeneralBookKey
-import net.bible.android.database.workspaces.WorkspaceEntities
 import net.bible.service.sword.SwordContentFacade
 import net.bible.service.sword.SwordDocumentFacade
 import org.crosswire.jsword.book.BookCategory
