+++ app/src/main/java/net/bible/android/control/page/CurrentPage.kt
@@ -20,13 +20,10 @@ package net.bible.android.control.page
 import android.app.Activity
 import android.view.Menu
 import net.bible.service.common.ParseException
-import net.bible.android.database.workspaces.WorkspaceEntities
 import net.bible.service.format.Note
 import org.crosswire.jsword.book.Book
 import org.crosswire.jsword.book.BookCategory
 import org.crosswire.jsword.passage.Key
-import org.json.JSONException
-import org.json.JSONObject
 
 /**
  * @author Martin Denham [mjdenham at gmail dot com]
