+++ app/src/main/java/net/bible/android/control/versification/ConvertibleVerse.kt
@@ -18,7 +18,7 @@
 package net.bible.android.control.versification
 
 import net.bible.android.control.page.ChapterVerse
-import net.bible.android.database.workspaces.WorkspaceEntities
+import net.bible.android.database.WorkspaceEntities
 import org.crosswire.jsword.passage.Verse
 import org.crosswire.jsword.versification.BibleBook
 import org.crosswire.jsword.versification.Versification
